Separate card deletion/creation into two functions
Breaks down the transferCards function into two functions: deleteCards & createCards. Creating the cards immediately after deletion caused the card to be added to the new project before the associated label had been removed from the relevant by the release-bot. If the two projects shared a label (ex. 17.07.1-ce-rc1 & 17.07.1-ce-rc2 share the label 17.07.1-ce/triage), the project card would be deleted from the new project because the label would be removed after it was created in the destination project. The label would then be added again and the card would be placed back into the second project. This would repeat multiple times for each project card and often end up deleting the project card entirely without adding it to the destination project.

By deleting all the project cards before creating their new counterparts in the destination project the release-bot has time to remove all the labels from the associated issues. This stops the multiple deletions, labels, and unlabels the release-bot was doing prior.

Signed-off-by: Corbin <corbin.coleman@docker.com>